Move Tunable classes to lynx_tunable file
